pyproject.toml
[project]
name = "heme-ovarian-pipeline"
version = "0.1.0"
description = "Heme metabolism ovarian cancer prognostic biomarker pipeline with multi-model consensus survival"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    "statsmodels>=0.14.0",
    "lifelines>=0.28.0",
    "scikit-survival>=0.23.0",
    "scikit-learn>=1.4.0",
    "xgboost>=2.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "gseapy>=1.1.0",
    "openpyxl>=3.1.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "matplotlib-venn>=0.11.0",
    "adjusttext>=0.8",
]

[project.optional-dependencies]
rpy2 = ["rpy2>=3.5.0"]
dev = ["pytest>=8.0.0", "ruff>=0.3.0", "mypy>=1.8.0"]

[project.scripts]
heme-pipeline = "heme_pipeline.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/heme_pipeline"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
